pyproject.toml
[project]
name = "langgraph-workspace"
version = "0.1.0"
description = "A workspace for LangGraph projects and tutorials"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dashscope>=1.24.5",
    "ipython>=9.5.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.1",
    "langchain-core>=0.1.0",
    "langchain-google-vertexai>=0.0.1",
    "langchain-openai>=0.3.33",
    "langgraph>=0.0.20",
    "langgraph-cli[inmem]>=0.4.2",
    "pyppeteer>=0.0.25",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
packages = [
    "common",
    "ice_break_agent",
    "langgraph_academy",
    "personality_agent",
    "other_script"
]

[tool.pytest.ini_options]
pythonpath = "."
